Task #46: Extend the Selenium test case base to better support testing against example modules.
Added a method to properly find and load a URL using a nodes UUID property - this is essential for using nodes loaded via YamlConfig as we do not know what their ID will be and many of them do not have consistent aliases. Also added a small class library for processing CSS colors to make checking for color equality easier.